updated code

+++ app/controllers/shops.js
@@ -1,68 +1,11 @@
 const Shop = require('../models/shops');
 const { validationResult } = require('express-validator');
-const { sendResponse, generateSlug, getEmailTemplate } = require('../helpers/utalityFunctions');
+const { sendResponse, generateSlug, getEmailTemplate,convertToObjectId } = require('../helpers/utalityFunctions');
 const messages = require('../messages/customMessages');
 const config = require("config");
 const sendEmail = require('../helpers/sendEmail');
 const User = require('../models/users')
-
-// Create Shop
-async function createShop(req, res) {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-        return res.status(400).json(sendResponse(1003, messages[1003], false, errors.array()));
-    }
-
-    try {
-        // Destructure keys from req.body
-        const { ownerName, address, phone, website, name, description, coverImage, logo, location } = req.body;
-
-        const existingShop = await Shop.findOne({ name, ownerId: req.loginUser._id });
-        if (existingShop) {
-            return res.status(400).json(sendResponse(1093, messages[1093], false));
-        }
-        const slug = generateSlug(name);
-        
-        // Check for valid geoLocation
-        let geoLocation;
-        if (location && location.lat && location.lng) {
-            geoLocation = {
-                type: 'Point',
-                coordinates: [location.lng, location.lat],
-                address: location.address,
-                zip: location.zip || '',          // Optional zip code
-                lat: location.lat,
-                lng: location.lng
-            };
-        } else {
-            return res.status(400).json(sendResponse(1000, "Invalid location data", false));
-        }
-
-        const newShop = new Shop({
-            ownerName,
-            ownerId: req.loginUser._id,
-            address,
-            phone,
-            website,
-            name,
-            slug,
-            description,
-            coverImage,
-            logo,
-            ...(geoLocation && { location: geoLocation })
-        });
-  
-        const shop = await newShop.save();
-        const userEmailData = { name: ownerName, shopName: name };
-        const userEmailHtml = await getEmailTemplate(userEmailData, "shopVerification.hbs", false, false);
-        await sendEmail(req.loginUser.email, config.mailEmail, userEmailHtml, userEmailData, "", "Your Shop is In Verification Process", "");
-
-        return res.status(201).json(sendResponse(1060, messages[1060], true, shop));
-
-    } catch (error) {
-        return res.status(500).json(sendResponse(1000, messages[1000], false, error.message));
-    }
-}
+const Order = require('../models/order')
 
 // Get All Shops
 async function getAllShops(req, res) {
@@ -73,11 +16,12 @@ async function getAllShops(req, res) {
 
     let options = {
         page: !req.query.page ? 1 : parseInt(req.query.page),
+        // select: { name: 1, email: 1, shopStatus: 1 },
         limit: !req.query.limit ? config.pageSize : parseInt(req.query.limit),
         sort: { createdAt: -1 },
     };
 
-    let match = { isApproved: true, isActive: true };
+    let match = {};
 
     try {
         const result = await Shop.paginate(match, options);
@@ -87,29 +31,25 @@ async function getAllShops(req, res) {
     }
 }
 
-// Get Pending Shops
-async function getPendingShops(req, res) {
+async function getAdminShop(req, res) {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
       return res.status(400).send(sendResponse(1003, messages[1003], false, errors.errors));
   }
 
-  // Set up pagination options
-  const options = {
-      page: parseInt(req.query.page) || 1, // Default to page 1
-      limit: parseInt(req.query.limit) || config.pageSize, // Default to configured page size
-      sort: { createdAt: -1 } // Optional: sort by createdAt or any other field
-  };
+  const adminId = convertToObjectId(req.loginUser._id);
+  console.log("Admin ID:", req.loginUser._id);
+  console.log("Converted Admin ID:", adminId);
 
   try {
-      // Use the paginate method to get paginated results
-      const result = await Shop.paginate({ isApproved: false }, options);
+      // Find the documents with the specified query, limit, and skip for pagination
+      const result = await Shop.findOne( { ownerId: adminId })
+      console.log("Query Result:", result); // Log the result for debugging
       return res.status(200).json(sendResponse(1064, messages[1064], true, result));
   } catch (err) {
       return res.status(500).json(sendResponse(1000, messages[1000], false, err.message));
   }
 }
-
 
 // Get Single Shop by ID
 async function getShopById(req, res) {
@@ -130,46 +70,79 @@ async function getShopById(req, res) {
 }
 
 // Approve Shop
-async function approveShop(req, res) {
+async function approveOrDisapproveShop(req, res) {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
       return res.status(400).send(sendResponse(1003, messages[1003], false, errors.errors));
   }
 
+  const { action, reason } = req.body; // Expecting { action: 'approve' } or { action: 'disapprove' }
+
   try {
+    
       // Find the shop first
       const shop = await Shop.findById(req.params.id);
       if (!shop) {
           return res.status(404).json(sendResponse(1065, messages[1065], false));
       }
-   // Check if the shop is already approved
-   if (shop.isApproved) {
-    return res.status(400).json(sendResponse(1097, messages[1097], false)); // Shop already approved
-  }
-      // Retrieve the owner's user information
-      const owner = await User.findById(shop.ownerId); // Assuming ownerId is the userId
-      if (!owner || !owner.email) {
-          return res.status(404).json(sendResponse(1005, messages[1005], false)); // Handle case where owner is not found or email is missing
+  
+      // Find the user associated with the shop's ownerId
+      const owner = await User.findById(shop.ownerId);
+      if (!owner) {
+          return res.status(404).json(sendResponse(1066, "Owner not found", false));
       }
 
-      // Update shop approval status
-      shop.isApproved = true;
-      shop.isActive = true;
+      // Check if the shop is associated with a user role and if the email is verified
+      if (shop.ownerId && owner.role === 2 && !owner.isEmailVerified) {
+          return res.status(403).send(sendResponse(1050, messages[1050], false, false)); // Inform about unverified user
+      }
 
-      const updatedShop = await shop.save();
+      if (action === 'approve') {
 
-      // Prepare and send email notification
-      const ownerEmail = owner.email;
-      const userEmailData = { name: owner.name, shopName: shop.name }; // Adjust this based on your user model
-      const userEmailHtml = await getEmailTemplate(userEmailData, "shopApproval.hbs", false, false);
-      
-      await sendEmail(ownerEmail, config.mailEmail, userEmailHtml, userEmailData, "", "Your Shop Has Been Approved", "");
+          shop.isActive = true;
+          shop.isApproved = true; // Set isApproved to true
+          owner.shopStatus = true; // Set shopStatus to true for the owner
 
-      return res.status(200).json(sendResponse(1096, messages[1096], true));
+          await Promise.all([shop.save(), owner.save()]); // Save both shop and owner
+
+          const agencyUserEmail = owner.email;
+          const emailData = {
+              name: shop.name,
+              agencyEmail: owner.email,
+              agencyUrl: `${config.agencyUrl}login`, // Unique login URL
+          };
+
+          const emailHtml = await getEmailTemplate(emailData, "agencyRegistration.hbs", false, false);
+          await sendEmail(agencyUserEmail, config.mailEmail, emailHtml, emailData, "", "Your Shop Has Been Approved", "");
+          return res.status(200).json(sendResponse(1096, messages[1096], true));
+      } 
+      else if (action === 'disapprove') {
+          shop.isActive = false;
+          shop.isApproved = false; // Set isApproved to false
+          owner.shopStatus = false; // Set shopStatus to false for the owner
+
+          await Promise.all([shop.save(), owner.save()]); // Save both shop and owner
+
+          // Prepare and send disapproval email notification
+          const agencyUserEmail = owner.email;
+          const emailData = {
+              name: shop.name,
+              agencyEmail: owner.email,
+              reason: reason
+          };
+
+          const emailHtml = await getEmailTemplate(emailData, "disapproveShop.hbs", false, false);
+          await sendEmail(agencyUserEmail, config.mailEmail, emailHtml, emailData, "", "Your Shop Has Been Disapproved", "");
+
+          return res.status(200).json(sendResponse(1099, "Your shop has been disapproved.", true));
+      } else {
+          return res.status(400).json(sendResponse(1004, "Invalid action specified.", false));
+      }
   } catch (error) {
       return res.status(500).json(sendResponse(1000, messages[1000], false, error.message));
   }
 }
+
 
 // Update Shop
 async function updateShop(req, res) {
@@ -231,24 +204,34 @@ async function deleteShop(req, res) {
     }
 
     try {
-        const shop = await Shop.findOne({_id:req.params.id, ownerId: req.loginUser._id });
+        // Find the shop by ID and check ownership
+        const shop = await Shop.findOne({ _id: req.params.id, ownerId: req.loginUser._id });
         if (!shop) {
             return res.status(404).json(sendResponse(1065, messages[1065], false));
         }
 
+        // Check if there are any non-completed orders for this shop
+        const pendingOrders = await Order.find({ shopId: req.params.id, orderStatus: { $ne: "completed" } });
+        if (pendingOrders.length > 0) {
+            return res.status(400).json(sendResponse(1066, "Cannot delete shop with pending orders.", false));
+        }
+
+        // Delete the shop if there are no non-completed orders
         await Shop.findByIdAndDelete(req.params.id);
-        return res.status(200).json(sendResponse(1098, messages[1098], true ));
+        await User.findByIdAndDelete(req.params.id);
+
+        return res.status(200).json(sendResponse(1098, messages[1098], true));
     } catch (error) {
+        console.error("Error deleting shop:", error);
         return res.status(500).json(sendResponse(1000, messages[1000], false, error.message));
     }
 }
 
 module.exports = {
-    createShop,
     getAllShops,
-    getPendingShops,
     getShopById,
-    approveShop,
+    approveOrDisapproveShop,
     updateShop,
-    deleteShop
+    deleteShop,
+    getAdminShop
 };
